Extract user functions and routes to own file
