+++ server/src/express-app.js
@@ -38,7 +38,8 @@ const cors = require('cors')
 const sep = path.sep;
 const tangyModules = require('./modules/index.js')()
 const {doesUserExist, extendSession, findUserByUsername, isSuperAdmin,
-  hashPassword, USERS_DB, login, getUserPermissions, updateUserSiteWidePermissions} = require('./auth')
+   USERS_DB, login, getUserPermissions, updateUserSiteWidePermissions} = require('./auth');
+const {registerUser,  getUserByUsername, isUserSuperAdmin, isUserAnAdminUser, getGroupsByUser} = require('./users');
 log.info('heartbeat')
 setInterval(() => log.info('heartbeat'), 5*60*1000)
 var cookieParser = require('cookie-parser');
@@ -244,54 +245,13 @@ app.get('/users/userExists/:username', isAuthenticated, async (req, res) => {
   }
 
 });
-app.post('/users/register-user', isAuthenticated, permit(['can_create_user']), async (req, res) => {
-  try {
-    if (!(await doesUserExist(req.body.username))) {
-      const user = req.body;
-      user.password = await hashPassword(user.password);
-      user.groups = [];
-      const data = await USERS_DB.post(user);
-      res.send({ statusCode: 200, data });
-      return data;
-    }
-  } catch (error) {
-    console.log(error);
-    return false; // @TODO return meaningful error
-  }
-});
+app.post('/users/register-user', isAuthenticated, permit(['can_create_user']), registerUser);
 
-app.get('/users/byUsername/:username', isAuthenticated, async (req, res) => {
-  const username = req.params.username;
-  try {
-    await USERS_DB.createIndex({ index: { fields: ['username'] } });
-    const results = await USERS_DB.find({ selector: { username: { '$regex': `(?i)${username}` } } });
-    const data = results.docs.map(username => username.username)
-    res.send({ data, statusCode: 200, statusMessage: 'Ok' });
-  } catch (error) {
-    console.error(error);
-    res.sendStatus(500);
-  }
-});
+app.get('/users/byUsername/:username', isAuthenticated, getUserByUsername);
 
-app.get('/users/isSuperAdminUser/:username', isAuthenticated, async (req, res) => {
-  try {
-    const data = await isSuperAdmin(req.params.username);
-    res.send({ data, statusCode: 200, statusMessage: 'ok' })
-  } catch (error) {
-    console.error(error);
-    res.sendStatus(500);
-  }
-});
+app.get('/users/isSuperAdminUser/:username', isAuthenticated, isUserSuperAdmin);
 
-app.get('/users/isAdminUser/:username', isAuthenticated, async (req, res) => {
-  try {
-    const data = await isAdminUser(req.params.username);
-    res.send({ data, statusCode: 200, statusMessage: 'ok' })
-  } catch (error) {
-    console.error(error);
-    res.sendStatus(500);
-  }
-});
+app.get('/users/isAdminUser/:username', isAuthenticated, isUserAnAdminUser);
 
 app.post('/editor/file/save', isAuthenticated, async function (req, res) {
   const filePath = req.body.filePath
@@ -333,53 +293,7 @@ app.get('/groups/:username', isAuthenticated, async function (req, res) {
     res.sendStatus(500)
   }
 })
-  
-async function getGroupsByUser(username) {
-  if (await isSuperAdmin(username)) {
-    const readdirPromisified = util.promisify(fs.readdir)
-    const files = await readdirPromisified('/tangerine/client/content/groups');
-    let filteredFiles = files.filter(junk.not).filter(name => name !== '.git' && name !== 'README.md')
-    let groups = [];
-    groups = filteredFiles.map((groupName) => {
-      return {
-        attributes: {
-          name: groupName,
-          role: 'admin'
-        }
-      }
-    })
-    return groups;
-  } else {
-    const user = await findUserByUsername(username);
-    let groups = [];
-    if (typeof user.groups !== 'undefined') {
-      groups = user.groups.map(group => {
-        return {
-          attributes: {
-            name: group.groupName,
-            role: group.role
-          }
-        }
-      });
-    }
-    return groups;
-  }
-}
-// If is not admin, return false else return the list of the groups to which user isAdmin
-async function isAdminUser(username) {
-  try {
-    const groups = await getGroupsByUser(username);
-    let data = groups.filter(group => group.attributes.role === 'admin');
-    if (data.length < 1) {
-      data = false;
-    }
-    return data;
-  }
-  catch (error) {
-    return false;
-    console.log(error)
-  }
-}
+
 app.post('/groups/:groupName/addUserToGroup', isAuthenticated, async (req, res) => {
   const payload = req.body;
   const groupName = req.params.groupName;
